@@ -50,7 +50,7 @@
         }
     },
     "require": {
-        "php": "^7.3 || ~8.0.0 || ~8.1.0",
+        "php": "^7.4 || ~8.0.0 || ~8.1.0",
         "ext-mongodb": "*",
         "container-interop/container-interop": "^1.2.0",
         "doctrine/doctrine-module": "^4.2.2",
@@ -116,6 +116,10 @@
         "sort-packages": true,
         "platform": {
             "ext-mongodb": "1.8.0"
+        },
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": true,
+            "composer/package-versions-deprecated": true
         }
     }
 }
